pyproject.toml
[project]
name = "auto-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "langchain[openai]>=1.0.7",
    "mcp[cli]>=1.21.0",
    "openai>=2.7.1",
    "pypdf>=6.2.0",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
]
